Refactor Installer.py for improved readability

+++ Installer.py
@@ -1,4 +1,4 @@
-import os 
+import os
 import sys
 import platform
 import tkinter as tk
@@ -7,6 +7,8 @@ basedir = os.getcwd()
 
 accessdev = ["Windows", "Linux"]
 filetointdir = ["L;", "Drivers", "System", "Lumeninate"]
+fileforl = ["documents"]
+filesforsys = ["LumenOsSystem", "System16", "Usersconfig", "bootemp", "NBFT", "Protocol", "FileWriting"]
 
 print(f"BaseDir Set as {basedir}")
 
@@ -15,19 +17,65 @@ root.title("Install LumenOS")
 root.geometry("450x450")
 delfile = None
 
+    
+
 def erad(a):
     global delfile
     # second installer step.
     print(a)
-    i = 1
-    while i < len(delfile):
-        if os.path.exists(delfile[i]):
-            os.remove(delfile[i])
-            i += 1
-        else:
-            raise RuntimeError("File not located")
-        
+    for widget in root.winfo_children():
+        widget.destroy()
+    
+    i = 0
+    welt = tk.Label(root, text="Deleting files")
+    weltper = tk.Label(root, text="0%")
+    welt.pack(pady=10)
+    weltper.pack(pady=10)
+    curdir = os.getcwd()
+    weltcur = tk.Label(root, text=curdir)
+    weltcur.pack(pady=10)
+    
+    
+    fulneed = 0
+    if len(delfile) > 0:
+        need = 100 / len(delfile)
+        while i < len(delfile):
+            if os.path.exists(delfile[i]):
+                os.remove(delfile[i])
+                fulneed += need
+                weltper.config(text = f"{fulneed}%")
+                root.update()
+                i += 1
+            else:
+                raise RuntimeError("File not located")
     print("Eradication complete!")
+    
+    global filetointdir
+    global fileforlglobal 
+    global filesforsys
+    
+    os.mkdir("LumenOs")
+    os.chdir("LumenOs")
+    
+    welt.config(text = "Creating Folders")
+    
+    curdir = os.getcwd()
+    weltcur = tk.Label(root, text=curdir)
+    
+    i = 0
+    need = 100/ len(filetointdir) 
+    fulneed = 0
+    i = 0
+    while i <  len(filetointdir):
+        os.mkdir(filetointdir[i])
+        fulneed += need
+        weltper.config(text = f"{fulneed}%")
+        root.update()
+        i+=1
+        
+    print("Main folder creation complete")
+    
+
 
 def installs1():
     global delfile
@@ -37,30 +85,40 @@ def installs1():
             print(file)
             delfile.append(file)
         else:
-            continue        
-    
+            continue
     for widget in root.winfo_children():
         widget.destroy()
-        
-    welt = tk.Label(root, text=f"The following files will be eradicated: {delfile} | {len(delfile)} File(s) will be eradicated.")
+    welt = tk.Label(
+        root,
+        text=f"The following files will be eradicated: {delfile} | {len(delfile)} File(s) will be eradicated.",
+    )
     welt.pack(pady=10)
-    
-    welt2 = tk.Button(root, text=f"Continue and delete files.", command=lambda: erad(delfile))
+
+    welt2 = tk.Button(
+        root, text=f"Install and delete files.", command=lambda: erad(delfile)
+    )
     welt2.pack(pady=10)
+
 
 if platform.system() in accessdev:
     print("Install allowed")
     welt = tk.Label(root, text="Welcome to the LumenOS Installer.")
     welt.pack(pady=10)
-    welt2 = tk.Label(root, text=f"Ensure the installer is in the desired installation folder. This folder must be empty, as all contents will be erased during installation.\n Current location: {basedir}" )
+    welt2 = tk.Label(
+        root,
+        text=f"Ensure the installer is in the desired installation folder. This folder must be empty, as all contents will be erased during installation.\n Current location: {basedir}",
+    )
     welt2.pack(pady=10)
-    
+
     welt3 = tk.Button(root, text="Install.", command=installs1)
     welt3.pack(pady=10)
-    
 else:
     print("Install disallowed")
-    welt = tk.Label(root, text="Your device is incompatible with LumenOS, THe installer will close shortly.")
+    welt = tk.Label(
+        root,
+        text="Your device is incompatible with LumenOS, THe installer will close shortly.",
+    )
     welt.pack(pady=10)
-    
+
     root.after(3000, root.destroy)  # closes after 3 seconds
+
